[692] Use a generic deleteTreeItem mutation for documents and objects
We can not use the same generic mutation for representation removal as
there is some special casing in EditingContextEventProcessor to
dispose deleted representations which assumes we use a
DeleteRepresentationInput for that.

Note that this change temporarily removes the confirmation step for
document deletion to align the behavior of the different tree item
kinds. This will be added back later in a more generic way.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/692
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>